requirements.txt
supabase==1.0.3
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
flask==3.0.2
flask-cors==4.0.0
